update callback and getPreviewApiKey to app router, remove pages router
